Extract the tests to upload photo into a separate file
